% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accum_test.r
\name{accum_test}
\alias{accum_test}
\title{Accumulation test with the hinge-exponential function}
\usage{
accum_test(x, alpha = 0.2, c = 2, strict = FALSE)
}
\arguments{
\item{x}{A vector of p-values to be tested IN ORDER.}

\item{alpha}{The desired FDR threshold.}

\item{c}{Parameter of the hinge-exponential function used in the test.}

\item{strict}{If TRUE, the procedure is more conservative but
controls the false discovery rate. If FALSE, the procedure is more liberal
but controls a modified version of the false discovery rate.}
}
\value{
A vector of rejections, has the form 1:k for some value of k.
}
\description{
Takes an ordered list of p-values and returns a set of selections
controlling the false discovery rate.
This is an implementation of the method of Li and Barber, JASA 2017.
}
