% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{col_cors}
\alias{col_cors}
\title{Correlation across columns}
\usage{
col_cors(preds, Y, genotypes = TRUE)
}
\arguments{
\item{preds}{Vector of length n OR a matrix of dimension (n x k) giving the logit inverse of the probability
of being one for k different models.}

\item{Y}{Observed responses, a 0-1 vector of length n or a matrix of dimension (n x k)}

\item{genotypes}{Whether to interpret the data as genotypes, i.e. summing pairs of rows
   of \code{logit_inv} gives the total logit inverse for the observation. 

Returns:}
}
\value{
A vector of length k, the correlations of the columns of Y
}
\description{
Returns correlation of a prediction direction(s) with an observed response(s).
  Intended for use as a feature-importance measure for the CRT.
}
