% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring_generation.r
\name{create_offspring_population}
\alias{create_offspring_population}
\title{Simulate an offspring population}
\usage{
create_offspring_population(parent_pop, d, n_offspring = 100,
  lambda = 0.012, scores = NULL, beta = 0)
}
\arguments{
\item{parent_pop}{n x p matrix of parental haplotypes}

\item{d}{genetic distance}

\item{n_offspring}{number of offspring to sample.}

\item{lambda}{recombination rate (recombinations per unit genetic distance)}

\item{scores}{n-vector of scores for preferred mating (e.g. inner product with first PC)}

\item{beta}{intensity of preferred mating.}
}
\value{
ancestry: n_offspring x 2 matrix giving parents for each sampled offspring

children: n_offspring x p matrix giving the offspring haplotypes.
}
\description{
Samples a population of offspring haplotypes with optional preferred mating.
}
