% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{deviance_test_stats}
\alias{deviance_test_stats}
\title{Computed test statistics based on deviance change for GLM}
\usage{
deviance_test_stats(pred_base, x, y, family = "gaussian")
}
\arguments{
\item{pred_base}{vector of baseline predictions}

\item{x}{matrix of new predictors}

\item{y}{vector of response observations}
}
\value{
A test statistic
}
\description{
Computed test statistics based on deviance change for GLM
}
