% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring_generation.r
\name{find_mate_pair}
\alias{find_mate_pair}
\title{Find a pair of haplotypes.}
\usage{
find_mate_pair(scores, beta = 1)
}
\arguments{
\item{scores:}{A vector of the lenght of the parent population. Gives the score for 
for each parent, such as the inner product with the 1st PC.}
}
\value{
a vector of length two giving the selected parents.
}
\description{
Returns a pair of haplotypes. A first haplotype is uniformly selected, and then 
  a second is sampled with higher probability assigned to similar haplotypes.
}
