% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_backward_}
\alias{forward_backward_}
\title{forward_backward_}
\usage{
forward_backward_(x, p1, p2, d, p, lambda, epsilon, start, end)
}
\arguments{
\item{x}{0-1 vector of length p giving the offspring haplotypes}

\item{p1}{0-1 vector of length p giving one parental haplotype}

\item{p2}{0-1 vector of length p giving the other parental haplotype}

\item{d}{numberic vector of length p - 1  giving the distance between each site}

\item{p}{length of the vector x}

\item{lambda}{the recombination rate, a multiplier for the distance d}

\item{epsilon}{the pointwise mutation rate}

\item{start}{indexes of the subset of x to perform the computation (1 indexed)}

\item{end}{indexes of the subset of x to perform the computation (1 indexed)}
}
\value{
p1_start: the probability that the latent state immediately before the specified window is 1,
       i.e. the offspring is being copied from p1.

p1_start_odd: probability of an odd number of recombinations given that 
      the latent state immediately before the specified window is 1.

p2_start_odd: probability of an odd number of recombinations given that 
      the latent state immediately before the specified window is 2.
}
\description{
Carry out the forward-backward algorithm for the HMM implied by knowing the two
parental haplotypes.
}
