% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{gen_id_to_haps}
\alias{gen_id_to_haps}
\title{Convert genotype IDs to haplotype IDs}
\usage{
gen_id_to_haps(ids)
}
\arguments{
\item{ids}{vector of ids (integers)}
}
\value{
Vector of ids, typically of size 2 * length(ids).
}
\description{
Convert genotype IDs to haplotype IDs
}
