% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring_generation.r
\name{generate_ancestry}
\alias{generate_ancestry}
\title{Generate ancestry information}
\usage{
generate_ancestry(d, lambda = 0.012)
}
\arguments{
\item{d}{Vector of genetic distances between sites. Should be the length of the haplotype.}

\item{lambda}{Recombination rate per unit of genetic distance.

Return value:}
}
\value{
A vector of length equal to x1 with each entry 0 or 1, 
   corresponding to the ancestry at each location.
}
\description{
Simulates recombination and returns vector of 0-1 indicating ancestry. 
Intended for intenal use with the \code{generate_offspring} function.
}
