% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{get_beta_glmnet}
\alias{get_beta_glmnet}
\title{Run CV glmnet}
\usage{
get_beta_glmnet(X, Y, train_idx, family = "gaussian", s = "lambda.min",
  parallel = FALSE)
}
\arguments{
\item{X}{and n by p matrix}

\item{Y}{a response vector of length n}

\item{train_idx}{a set of indices to train on}

\item{family}{type of regression, either "gaussian" or "binomial"}

\item{s}{coefficient vector to  return. either "lambda.min" or "lambda.1se"}

\item{parallel}{set to TRUE to run CV in parallel. Must initialize DoMC or DoParallel
prior to execution.}
}
\value{
Fitted coefficients, a vector of length p + 1.
     First coordinate is the intercept.
}
\description{
Returns the fitted coefficients from a lasso fit after tuning the
  regularization parameter by cross-validation. Intended for internal
  use with the \code{linear_crt} function.
}
