% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_helpers.r
\name{get_fb_results}
\alias{get_fb_results}
\title{Forward-backward algorithm for parent-offspring trio of haplotypes}
\usage{
get_fb_results(H, H_parent, anc, group, d, lambda = 0.012,
  epsilon = 1e-08, window_size = 200)
}
\arguments{
\item{H}{an (n x p) matrix of the subject haplotypes}

\item{H_parent}{and (n_2 x p) matrix of the parental haplotypes}

\item{anc}{a (n x 2) table of ancestries.
anc[i, 1] and and[i, 2] give the rows of H_parent corresponding to the
parents of subject i}

\item{group}{a list of of indices of the group to test. Should be a continuous region,
such as 10:20.}

\item{d}{a vector of length p giving the genetic distances between sites}

\item{lambda}{the recombination rate, in units of d}

\item{epsilon}{the pointwise mutation rate}

\item{window_size}{the number of sites to on each size of the group to
use when computing the forward-backward probabilities. Smaller groups improve speed
and avoid numerical issues.}
}
\value{
a (n x 3) matrix giving the results of the forward-backward algorithm.
See the documentation of the \code{forward_backward_} function for output information.
}
\description{
Returns the result of the forward backward algorithm for a sepcified group.
  Intended for use as a pre-computation step for the \code{linear_crt} function.
}
