% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{get_genotype_ids}
\alias{get_genotype_ids}
\title{Complete haplotype list}
\usage{
get_genotype_ids(ids)
}
\arguments{
\item{ids}{Vector of ids (integers)}
}
\value{
Vector of ids, typically of size 2 * length(ids)
}
\description{
Returns a vector of all haplotypes from a set of individuals, given
  a list of some of the haplotypes from those individuals.
}
