% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasing_alignment.r
\name{get_phased_ancestry}
\alias{get_phased_ancestry}
\title{Indentify maternal vs paternal haplotypes}
\usage{
get_phased_ancestry(sample, haps_matrix)
}
\arguments{
\item{sample}{A dataframe in .sample format. \code{sample$father} gives the index of
the father in the \code{haps_matrix}, and the \code{sample$mother} gives the index of the 
the mother in the \code{haps_matrix}.}

\item{haps_matrix}{A n x p matrix of 0-1. Each row represents one haplotypes. 
To convert from indices to haplotype indices, an entry i in \code{sample$father}
corresponds to the two rows 2*i - 1 and 2*i in \code{haps_matrix}.}
}
\value{
A data frame with the following elements:

subject: the row ID of the subject. An entry of i corresponds to
 rows 2*i -1 and 2*i of \code{haps_matrix}.

chromosome: the chromosome number 1-22.

ancestor_1: the ID of the ancestor corresponding to the first haplotype
     of the offspring in \code{haps_matrix}.

ancestor_2: the ID of the ancestor corresponding to the second haplotype
     of the offspring in \code{haps_matrix}.
}
\description{
Identifies which of a subjects two haplotyes
is the maternal haplotype and which is the paternal haplotype.
}
