% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{haps_id_to_gen}
\alias{haps_id_to_gen}
\title{Convert haplotype IDs to genptype IDs}
\usage{
haps_id_to_gen(ids)
}
\arguments{
\item{ids}{vector of ids (integers)}
}
\value{
Vector of ids, typically of size length(ids) / 2
}
\description{
Convert haplotype IDs to genptype IDs
}
