% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haps_matrix}
\alias{haps_matrix}
\title{3,000 synthetic haplotypes from one chromosome}
\format{A matrix comprised of 0-1 entries with 3000 rows and 500 columns.}
\usage{
haps_matrix
}
\description{
A dataset containing 3,000 synthetic haplotypes of 500 features. 
Each haplotype is simulated from a fastPHASE-type hidden Markov model.
Each row represents and independent haplotype, and each column represents a genetic variant.
}
\keyword{datasets}
