% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{haps_to_gen}
\alias{haps_to_gen}
\title{Convert haplotype matrix to genotype matrix}
\usage{
haps_to_gen(H)
}
\arguments{
\item{H}{An array with an even length in the first dimension.}
}
\value{
An array of the same dimension as H, with the first dimension half as long.
}
\description{
Adds pairs of rows of a matrix to convert a haplotype matrix
to a genotype matrix. For example, rows 1 and 2 are added to create
row 1 of the output matrix. Rows 3 and 4 are added to create row 2
of the output matrix, and so on.
}
\details{
This function can also handle general arrays. It assumes that the 
first dimension indexes the haplotypes.
}
