% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accum_test.r
\name{hinge_exp}
\alias{hinge_exp}
\title{Hinge-exponential function}
\usage{
hinge_exp(x, c = 2)
}
\arguments{
\item{x}{A vector of p-values.}

\item{c}{A parameter of the accumulation test.}
}
\value{
A vector with the same length as x of 
   the element-wise application of 
		the hinge-exponential function.
}
\description{
Intended for internal use with the \code{accum_test} function.
}
