% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_helpers.r
\name{impute_variant_fast}
\alias{impute_variant_fast}
\title{Imputes a set of variants}
\usage{
impute_variant_fast(x, p1, p2, group, d, fb_result = NULL,
  lambda = 0.012, epsilon = 1e-08, window_size = 200)
}
\arguments{
\item{x}{0-1 vector of length p. The observed haplotypes of the offspring.}

\item{p1}{0-1 vector of length p. One of the parental haplotypes 
giving rise to \code{x}.}

\item{p2}{0-1 vector of length p. The other parental haplotypes 
giving rise to \code{x}.}

\item{group}{a subset of indices of 1:p. This subset of variants will be imputed.}

\item{d}{numeric vector of length p. The genetic distance between variants}

\item{lambda}{the recombination rate, in units of the vector d}

\item{epsilon}{the pointwise mutation rate}

\item{window_size}{when imputing the variants, look only at a window of this size
surrounding the group g. A smaller window size leads to faster computation
and avoids numerical issues.}
}
\value{
A vector of the same length as \code{group} containing the imputed variants.
}
\description{
Imputes the variants in a specified group, conditional on other variants,
 based on the recombination model. This is intented for use a subroutine of
  the conditional randomization test.
}
