% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_crt.r
\name{logistic_ll}
\alias{logistic_ll}
\title{Logistic-log likelihood}
\usage{
logistic_ll(logit_inv, Y, genotypes = TRUE)
}
\arguments{
\item{logit_inv}{Aector of length n OR a matrix of dimension (n x k) giving the logit inverse of the probability
of being one for k different models.}

\item{Y}{Observed responses, a 0-1 vector of length n or a matrix of dimension (n x k).}

\item{genotypes}{Whether to interpret the data as genotypes, i.e. summing pairs of rows
of \code{logit_inv} gives the total logit inverse for the observation.}
}
\value{
A vector of length k, the log-likelihoods of the columns of Y
}
\description{
Returns the log-likelihood of logistic regression fit(s). Intended for use
  as a feature-importance measure for the CRT.
}
