% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_ancestry_}
\alias{sample_ancestry_}
\title{sample_ancestry_}
\usage{
sample_ancestry_(x, p1, p2, d, p, lambda, epsilon)
}
\arguments{
\item{x}{0-1 vector of length p giving the offspring haplotypes}

\item{p1}{0-1 vector of length p giving one parental haplotype}

\item{p2}{0-1 vector of length p giving the other parental haplotype}

\item{d}{numberic vector of length p - 1  giving the distance between each site}

\item{p}{length of the vector x}

\item{lambda}{the recombination rate, a multiplier for the distance d}

\item{epsilon}{the pointwise mutation rate}
}
\value{
a vector of of the ancestral states
}
\description{
Sample the latent state of ancenstry indicators for a haplotype given
  its two ancestral haplotypes.
}
