% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt_helpers.r
\name{sample_parity_poisson}
\alias{sample_parity_poisson}
\title{Sample poisson conditional on its parity}
\usage{
sample_parity_poisson(lambda, parity = "even")
}
\arguments{
\item{lambda}{Poisson parameter.}

\item{parity}{Either "even" or "odd".}
}
\value{
Scalar, a sample from the poisson conditional on the specified parity
}
\description{
Sample a poisson conditional on it being even or odd
  This function is only computationally efficient when lambda is small.
}
