% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdt.r
\name{tdt}
\alias{tdt}
\title{Computes TDT p-value}
\usage{
tdt(y, child, p1, p2)
}
\arguments{
\item{y}{0-1 vector, outcome status}

\item{child}{0-1 vector, haplotype of child}

\item{p1}{0-1 vector, haplotype of one parent}

\item{p2}{0-1 vector, haplotype of second parent}
}
\value{
p-value
}
\description{
Arguments:
}
